\name{cggmisc.mccullochquantile}
\alias{mccullochquantile}
\title{McCulloch's Quantile Estimator of the Parameters of an Alpha-Stable Distribution}

\description{
	Returns 
}

\usage{
}

\arguments{

  \item{ x }{ A numeric vector of data. Missing values (NA) are allowed. } 
}


\details{
}

\value{

}
\author{ Ported to R by Christopher G. Green \email{christopher.g.green@gmail.com}.
}
\references{
}
\examples{

}
\keyword{alpha-stable}
